nes
Fixes for vbl_clear_time.nes and power_up_palette.nes.  Fixed 2002.7 to be properly cleared when 2002 is read.  Made the vblank signal last the appropriate amount of time to pass vbl_clear_time.nes.

Note, I had a problem with BurgerTime at first, so I "fixed" a bug that exists in the real NES.  If 2002 is read exactly when VBL goes high, it will return 0 and clear the flag.  This probably has something to do with timing that I will have to come back to later.

I also fixed up the palette mirroring to fix a bug in ExciteBike.

[git-p4: depot-paths = "//depot/hw/nes/": change = 163]